requirements.txt
charmhelpers
charms.reactive

